+++ mago3d-core/src/doc/ja/database/ddl/dataInfo_log.sql
@@ -3,7 +3,7 @@ drop table if exists data_info_log cascade;
 
 create table data_info_log(
 	data_info_log_id				bigint,
-	project_id						smallint,
+	project_id						int,
 	data_id							bigint,
 	user_id							varchar(32),
 	latitude						numeric(13,10),
