+++ mago3d-core/src/doc/ko/database/ddl/widget.sql
@@ -2,9 +2,9 @@ drop table if exists widget cascade;
 
 -- À§Á¬
 create table widget(
-	widget_id			smallint,
+	widget_id			int,
 	name				varchar(100)			not null ,
-	view_order			smallint								default '1',
+	view_order			int								default '1',
 	user_id				varchar(32)	 			not null,
 	insert_date				timestamp without time zone			default now(),
 	constraint widget_pk primary key (widget_id)
